Review modifiche end point get Weather forecaste e creazione end point  per inserirne di nuove
