+++ WeatherForecast.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33414.496
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApplication2", "WebApplication2\WebApplication2.csproj", "{B369B91C-5485-4C78-A723-9C73AA9B2C1E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WatherForecats", "WeatherForecast\WatherForecats.csproj", "{B369B91C-5485-4C78-A723-9C73AA9B2C1E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
