+++ WeatherForecast/WeatherForecast.cs
@@ -1,0 +1,13 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace WeatherForecast
+{
+    public class WeatherForecast
+    {
+        public DateTime Date { get; set; }
+        [Required]
+        public int? TemperatureC { get; set; }
+
+        public string? Summary { get; set; }
+    }
+}
